# src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "My mathaliapp API"
  },
  "host" : "127.0.0.1:8080",
  "basePath" : "/",
  "tags" : [ {
    "name" : "User Management System",
    "description" : "Operations for users Management System"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/api/private/users/add" : {
      "post" : {
        "tags" : [ "User Management System" ],
        "summary" : "add user ",
        "description" : "",
        "operationId" : "addUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/private/users/all" : {
      "get" : {
        "tags" : [ "User Management System" ],
        "summary" : "Get users",
        "description" : "",
        "operationId" : "getAllUsers",
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved list",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          },
          "401" : {
            "description" : "You are not authorized to view the resource"
          },
          "403" : {
            "description" : "Accessing the resource you were trying to reach is forbidden"
          },
          "404" : {
            "description" : "The resource you were trying to reach is not found"
          }
        }
      }
    },
    "/api/private/users/delete/{id}" : {
      "delete" : {
        "tags" : [ "User Management System" ],
        "summary" : "delete user by Id",
        "description" : "",
        "operationId" : "deleteUserById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/private/users/deleteByUserIdent/{userIdent}" : {
      "delete" : {
        "tags" : [ "User Management System" ],
        "summary" : "delete user by Id",
        "description" : "",
        "operationId" : "deleteUserByUserIdent",
        "parameters" : [ {
          "name" : "userIdent",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/private/users/find/{id}" : {
      "get" : {
        "tags" : [ "User Management System" ],
        "summary" : "Get user by Id",
        "description" : "",
        "operationId" : "getUserById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/private/users/update" : {
      "put" : {
        "tags" : [ "User Management System" ],
        "summary" : "update user ",
        "description" : "",
        "operationId" : "updateUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Role" : {
      "type" : "object",
      "required" : [ "libelle" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "libelle" : {
          "type" : "string"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "firstname" : {
          "type" : "string"
        },
        "lastname" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "phone" : {
          "type" : "string"
        },
        "imageUrl" : {
          "type" : "string"
        },
        "userIdent" : {
          "type" : "string"
        },
        "roles" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/Role"
          }
        }
      }
    }
  }
}